We make the prompt aware of extruder temperature.
In order to do this, we consolidate handeling of printer status
with a Status class. The status class is updated by recvcb.

This has the side effect of simplifying the implementation of gettemp.

We also detect whether there is a heated build platform or not, and
don't display info about it if there isn't.