Rework temperature report parsing
This change is based on an IRC discussion with @Triffid_Hunder, @Sound and
reifsnyder, where we figured out that the best way to report multiple extruders
temperature would simply to add more information to the M105 line in the form of
T0:a/b T1:a/b... The "T:" part should still report the active extruder tool by
default, or obey the T gcode parameter).
This change also simplifies the parsing by using a generic regexp.