Changed extrusion length to hidden setting
Pulled latest master from kliment/printun and changed extrusion length
to a hidden setting. It is now saved upon change as
settings.last_extrusion and can be set as such. Also added default
extrusion length that may be used as a fall back for deploying printrun
with a different default extrusion length without having to change that
everywhere in the code.