Resume print from selected layer #383
Implement resuming a print from a user selected layer.
There is no requirement that the print has been paused or started
in the current pronterface session.
The command is accessed through menu Advanced -> Print from layer.
The layer selection is done with the slider in the standalone 3D view
window, which is opened on double click in the embeded 3D viewer.
The printer is initialized by filtering some commands in the skipped
lines.
The last (top) visible layer in the stanalone 3D viewer will be printed,
so you have to slide to the lower layer to compare visually the printed
object with the 3D visualization and then slide up to select the to be
printed layer.
The Z value in the 3D viewer status should be the same as the first sent
gcode commands.
The layer numbering in this commit does not coincide with the gcode
comments, fixed in the next commit.
The starting commands are shown for confirmation.
The code is not tested on real printer, just with
 ./testtools/mock-printer.