p/printcore: Refactoring split into device.py
Split `printcore.py` into two. All the functionality related to
directly communicating with serial ports or TCP sockets is moved into
`device.py`. This way this logic is abstracted from `printcore.py`,
which is now independent of the underlying connection type.

This reduces the complexity of `printcore.py` and increases code
modularity. Plus it has the advantage of making it simple to add more
connection types in the future (if any).