Move static metadata to pyproject.toml
The philosophy being to keep all static project metadata within
'pyproject.toml' while all build configuration and dynamic metadata
remains at 'setup.py'.

The 'long_description' is no longer needed because we link to the README
on 'pyproject.toml'.

A 'license' field is not required because it is already set on
'classifiers'.