refactor: Remove legacy custombtn and configfile
custombtn.txt was deprecated before Printrun 2.x but the code was left. The
code didn't even work as expected so this potentially breaking change has a
very low chance of affecting any users.
Closes #1236

The `configfile` function was outdated and only used for handling
custombtn.txt so it is safe to remove as well.
Closes #1286